+++ src/FLQuant_base.cpp
@@ -636,8 +636,8 @@ FLQuant_base<T>& FLQuant_base<T>::operator *= (const FLQuant_base<T2>& rhs){
 // FLQuantAdolc / CppAD *= adouble
 template <typename T>
 FLQuant_base<T>& FLQuant_base<T>::operator *= (const T& rhs){
-    //Rprintf("In scalar T=*T multiplication assignment\n");
-    std::transform((*this).data.begin(), (*this).data.end(), (*this).data.begin(), std::bind1st(std::multiplies<T>(),rhs)); 
+    std::transform((*this).data.begin(), (*this).data.end(), (*this).data.begin(),
+        [&rhs](const T& val) { return val * rhs; });
     return *this;
 }
 
@@ -647,8 +647,8 @@ FLQuant_base<T>& FLQuant_base<T>::operator *= (const T& rhs){
 template <typename T>
 template <typename T2>
 FLQuant_base<T>& FLQuant_base<T>::operator *= (const T2& rhs){
-    //Rprintf("In scalar T=*T2 multiplication assignment\n");
-    std::transform((*this).data.begin(), (*this).data.end(), (*this).data.begin(), std::bind1st(std::multiplies<T>(),rhs)); 
+    std::transform((*this).data.begin(), (*this).data.end(), (*this).data.begin(),
+        [&rhs](const T& val) { return val * static_cast<T>(rhs); });
     return *this;
 }
 
@@ -834,7 +834,8 @@ template <typename T>
 FLQuant_base<T> operator / (const T& lhs, const FLQuant_base<T>& rhs){
     FLQuant_base<T> out = rhs;
     std::vector<T> out_data = out.get_data();
-    std::transform(out_data.begin(), out_data.end(), out_data.begin(), std::bind1st(std::divides<T>(),lhs)); 
+    std::transform(out_data.begin(), out_data.end(), out_data.begin(),
+        [&lhs](const T& val) { return lhs / val; });
     out.set_data(out_data);
     return out;
 }
@@ -843,7 +844,8 @@ template <typename T>
 FLQuant_base<T> operator / (const double& lhs, const FLQuant_base<T>& rhs){
     FLQuant_base<T> out = rhs;
     std::vector<T> out_data = out.get_data();
-    std::transform(out_data.begin(), out_data.end(), out_data.begin(), std::bind1st(std::divides<T>(),lhs)); 
+    std::transform(out_data.begin(), out_data.end(), out_data.begin(),
+        [&lhs](const T& val) { return static_cast<T>(lhs) / val; });
     out.set_data(out_data);
     return out;
 }
@@ -851,10 +853,12 @@ FLQuant_base<T> operator / (const double& lhs, const FLQuant_base<T>& rhs){
 FLQuant_base<double> operator / (const double& lhs, const FLQuant_base<double>& rhs){
     FLQuant_base<double> out = rhs;
     std::vector<double> out_data = out.get_data();
-    std::transform(out_data.begin(), out_data.end(), out_data.begin(), std::bind1st(std::divides<double>(),lhs)); 
+    std::transform(out_data.begin(), out_data.end(), out_data.begin(),
+        [&lhs](const double& val) { return lhs / val; });
     out.set_data(out_data);
     return out;
 }
+
 
 template <typename T>
 FLQuant_base<T> operator / (const FLQuant_base<T>& lhs, const double& rhs){
@@ -874,7 +878,8 @@ template <typename T>
 FLQuant_base<T> operator / (const T& lhs, const FLQuant_base<double>& rhs){
     FLQuant_base<T> out(rhs);
     std::vector<T> out_data = out.get_data();
-    std::transform(out_data.begin(), out_data.end(), out_data.begin(), std::bind1st(std::divides<T>(),lhs)); 
+    std::transform(out_data.begin(), out_data.end(), out_data.begin(),
+        [&lhs](const T& val) { return lhs / val; });
     out.set_data(out_data);
     return out;
 }
@@ -925,17 +930,18 @@ FLQuant_base<T>& FLQuant_base<T>::operator -= (const FLQuant_base<T2>& rhs){
 // FLQuantAdolc / CppAD -= adouble
 template <typename T>
 FLQuant_base<T>& FLQuant_base<T>::operator -= (const T& rhs){
-    std::transform((*this).data.begin(), (*this).data.end(), (*this).data.begin(), std::bind2nd(std::minus<T>(),rhs)); 
+    std::transform((*this).data.begin(), (*this).data.end(), (*this).data.begin(),
+        [&rhs](const T& val) { return val - rhs; });
     return *this;
 }
-
 // Special case
 // Used for FLQuantAdolc / CppAD -= double
 // Needs to be instanitated due to extra template class, T2
 template <typename T>
 template <typename T2>
 FLQuant_base<T>& FLQuant_base<T>::operator -= (const T2& rhs){
-    std::transform((*this).data.begin(), (*this).data.end(), (*this).data.begin(), std::bind2nd(std::minus<T>(),rhs)); 
+    std::transform((*this).data.begin(), (*this).data.end(), (*this).data.begin(),
+        [&rhs](const T& val) { return val - static_cast<T>(rhs); });
     return *this;
 }
 
@@ -975,7 +981,8 @@ template <typename T>
 FLQuant_base<T> operator - (const T& lhs, const FLQuant_base<T>& rhs){
     FLQuant_base<T> out = rhs;
     std::vector<T> out_data = out.get_data();
-    std::transform(out_data.begin(), out_data.end(), out_data.begin(), std::bind1st(std::minus<T>(),lhs)); 
+    std::transform(out_data.begin(), out_data.end(), out_data.begin(),
+        [&lhs](const T& val) { return lhs - val; });
     out.set_data(out_data);
     return out;
 }
@@ -984,7 +991,8 @@ template <typename T>
 FLQuant_base<T> operator - (const double& lhs, const FLQuant_base<T>& rhs){
     FLQuant_base<T> out = rhs;
     std::vector<T> out_data = out.get_data();
-    std::transform(out_data.begin(), out_data.end(), out_data.begin(), std::bind1st(std::minus<T>(),lhs)); 
+    std::transform(out_data.begin(), out_data.end(), out_data.begin(),
+        [&lhs](const T& val) { return static_cast<T>(lhs) - val; });
     out.set_data(out_data);
     return out;
 }
@@ -992,7 +1000,8 @@ FLQuant_base<T> operator - (const double& lhs, const FLQuant_base<T>& rhs){
 FLQuant_base<double> operator - (const double& lhs, const FLQuant_base<double>& rhs){
     FLQuant_base<double> out = rhs;
     std::vector<double> out_data = out.get_data();
-    std::transform(out_data.begin(), out_data.end(), out_data.begin(), std::bind1st(std::minus<double>(),lhs)); 
+    std::transform(out_data.begin(), out_data.end(), out_data.begin(),
+        [&lhs](const double& val) { return lhs - val; });
     out.set_data(out_data);
     return out;
 }
@@ -1015,7 +1024,8 @@ template <typename T>
 FLQuant_base<T> operator - (const T& lhs, const FLQuant_base<double>& rhs){
     FLQuant_base<T> out(rhs);
     std::vector<T> out_data = out.get_data();
-    std::transform(out_data.begin(), out_data.end(), out_data.begin(), std::bind1st(std::minus<T>(),lhs)); 
+    std::transform(out_data.begin(), out_data.end(), out_data.begin(),
+        [&lhs](const T& val) { return lhs - val; });
     out.set_data(out_data);
     return out;
 }
@@ -1068,7 +1078,8 @@ FLQuant_base<T>& FLQuant_base<T>::operator += (const FLQuant_base<T2>& rhs){
 // FLQuantAdolc / CppAD += adouble
 template <typename T>
 FLQuant_base<T>& FLQuant_base<T>::operator += (const T& rhs){
-    std::transform((*this).data.begin(), (*this).data.end(), (*this).data.begin(), std::bind1st(std::plus<T>(),rhs)); 
+    std::transform((*this).data.begin(), (*this).data.end(), (*this).data.begin(),
+        [&rhs](const T& val) { return val + rhs; });
     return *this;
 }
 
@@ -1078,7 +1089,8 @@ FLQuant_base<T>& FLQuant_base<T>::operator += (const T& rhs){
 template <typename T>
 template <typename T2>
 FLQuant_base<T>& FLQuant_base<T>::operator += (const T2& rhs){
-    std::transform((*this).data.begin(), (*this).data.end(), (*this).data.begin(), std::bind1st(std::plus<T>(),rhs)); 
+    std::transform((*this).data.begin(), (*this).data.end(), (*this).data.begin(),
+        [&rhs](const T& val) { return val + static_cast<T>(rhs); });
     return *this;
 }